create_test_data: move to spinalcordtoolbox.testing
Solves issue #2636
