+++ spinalcordtoolbox/testing/create_test_data.py
